fix size_t int

+++ tasks/omp/kapustin_i_jarv_alg/include/ops_omp.hpp
@@ -22,7 +22,7 @@ class TestTaskOMP : public ppc::core::Task {
   std::vector<std::pair<int, int>> output_;
   std::pair<int, int> current_point_;
   std::pair<int, int> next_point_;
-  size_t leftmost_index_;
+  int leftmost_index_;
 
   [[nodiscard]] static int Orientation(const std::pair<int, int>& p, const std::pair<int, int>& q,
                                        const std::pair<int, int>& r);
